19916: Changes expf to exp with type to address gcc bug prior to gcc 13

+++ src/Amalgam/GeneralizedDistance.h
@@ -259,7 +259,7 @@ public:
 		{
 			//multiplying by the reciprocal is lower accuracy due to rounding differences but faster
 			double deviation_reciprocal = feature_attribs.deviationReciprocal;
-			diff += std::expf(static_cast<float>(-diff * deviation_reciprocal)) * (3 * deviation + diff) * 0.5;
+			diff += std::exp(static_cast<float>(-diff * deviation_reciprocal)) * (3 * deviation + diff) * 0.5;
 			if(!surprisal_transform)
 				return diff;
 			else
@@ -279,7 +279,7 @@ public:
 		else //!high_accuracy
 		{
 			//2*sigma*(e^(-1*(diff^2)/((2*simga)^2)))/sqrt(pi) - diff*erfc(diff/(2*sigma))
-			diff += s_two_over_sqrt_pi * deviation * std::expf(static_cast<float>(-term * term)) - diff * std::erfc(term);
+			diff += s_two_over_sqrt_pi * deviation * std::exp(static_cast<float>(-term * term)) - diff * std::erfc(term);
 			if(!surprisal_transform)
 				return diff;
 			else
